+++ src/app/components/oneprovider-view-container.js
@@ -56,8 +56,9 @@ export default Component.extend(I18n, ChooseDefaultOneprovider, {
   pointerEvents: service(),
 
   /**
+   * View to show in legacy Oneprovider, eg.spaces, transfers, shares
    * @virtual optional
-   * View to show in legacy Oneprovider, eg. spaces, transfers, shares
+   * @deprecated TODO: VFS-7573 remove support for v19 Oneproviders
    * @type {String}
    */
   resourceType: '',
@@ -298,7 +299,11 @@ export default Component.extend(I18n, ChooseDefaultOneprovider, {
   selectDefaultProvider(providers = this.get('providers')) {
     return this.chooseDefaultOneprovider(providers).then(defaultProvider => {
       if (defaultProvider) {
-        this.get('oneproviderIdChanged')(get(defaultProvider, 'entityId'));
+        const providerId = get(defaultProvider, 'entityId');
+        this.get('oneproviderIdChanged')(
+          providerId,
+          true,
+        );
       }
     });
   },