+++ src/app/components/sidebar-spaces/second-level-items.js
@@ -103,6 +103,14 @@ export default SecondLevelItems.extend(I18n, {
     }
   ),
 
+  itemDatasets: computed(function itemDatasets() {
+    return {
+      id: 'datasets',
+      label: this.t('aspects.datasets'),
+      icon: 'browser-dataset',
+    };
+  }),
+
   itemProviders: computed(function itemProviders() {
     return {
       id: 'providers',
@@ -167,6 +175,7 @@ export default SecondLevelItems.extend(I18n, {
     'itemData',
     'itemShares',
     'itemTransfers',
+    'itemDatasets',
     'itemProviders',
     'itemMembers',
     'itemHarvesters',