+++ src/app/locales/en/components/errors/share-not-found.js
@@ -1,0 +1,4 @@
+export default {
+  shareNotFound: 'Share not found',
+  shareNotFoundDescription: 'This share points to an inexistent file/directory. Most likely it has been deleted.',
+};