+++ src/app/locales/en/components/sidebar-groups.js
@@ -10,5 +10,7 @@ export default {
     leave: 'Leave',
     remove: 'Remove',
     groupPersistence: 'group persistence',
+    copyId: 'Copy group ID',
+    groupId: 'group ID',
   },
 };