+++ src/app/locales/en/components/sidebar-spaces.js
@@ -2,6 +2,8 @@ export default {
   spaceItem: {
     rename: 'Rename',
     leave: 'Leave',
+    copyId: 'Copy space ID',
+    spaceId: 'space ID',
     totalStorageHint: 'Total size of the space: {{{size}}}',
     supportCountHint: 'Number of supporting providers: {{{count}}}',
     spacePersistence: 'space persistence',
@@ -9,8 +11,9 @@ export default {
   secondLevelItems: {
     aspects: {
       index: 'Overview',
-      data: 'Data',
+      data: 'Files',
       transfers: 'Transfers',
+      datasets: 'Datasets',
       shares: 'Shares',
       providers: 'Providers',
       members: 'Members',